리스트 삭제, 수정, 셔플 기능 추가

+++ src/components/FoodItem.tsx
@@ -1,8 +1,72 @@
 import React, { useState } from 'react';
-const FoodItem = (props: { id: number; name: string, deleteItem: (id: number) => void; }) => {
-  const { id, deleteItem } = props;
-  return <div>{props.name}<button type="button" onClick={() => {
-    deleteItem(id);
-  }}>delete</button></div>;
+
+const FoodItem = (props: {
+  id: number;
+  name: string;
+  select: string;
+  updataItem: (id: number, name: string) => void;
+  deleteItem: (id: number) => void;
+}) => {
+  const { id, select, updataItem, deleteItem } = props;
+  const [editing, setEditing] = useState(false);
+  const [name, setName] = useState(props.name);
+
+  if (editing) {
+    return (
+      <div className={select}>
+        <input
+          type="text"
+          value={name}
+          onChange={(e) => {
+            setName(e.target.value);
+          }}
+        />
+        <button
+          type="button"
+          className="apply"
+          onClick={() => {
+            setEditing(!editing);
+            updataItem(id, name);
+          }}
+        >
+          apply
+        </button>
+        <button
+          type="button"
+          className="delete"
+          onClick={() => {
+            deleteItem(id);
+          }}
+        >
+          delete
+        </button>
+      </div>
+    );
+  }
+
+  return (
+    <div className={select}>
+      {name}
+      <button
+        type="button"
+        className="modify"
+        onClick={() => {
+          setEditing(!editing);
+        }}
+      >
+        modify
+      </button>
+      <button
+        type="button"
+        className="delete"
+        onClick={() => {
+          deleteItem(id);
+        }}
+      >
+        delete
+      </button>
+    </div>
+  );
 };
+
 export default FoodItem;
